bugzilla.py: allow the user change the base bugzilla url
In order to allow the user to change which bugzilla the lib will fetch
the data, the method get_django_bugzilla has been updated to use the url
from the variable BUGZILLA_URL defined in the Django settings. If the
variable is not defined, the default value will be used instead.

This is useful because help application developer which use this lib
define a test bugzilla instance easily.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>